ml-service/requirements.txt
sentence-transformers==2.5.0
faiss-cpu==1.8.0
fastapi==0.110.1
uvicorn==0.25.0
pydantic>=2.6.4
python-dotenv>=1.0.1
numpy>=1.26.0
openai>=1.12.0
httpx>=0.27.0
